requirements.txt
APScheduler==3.6.3
asgiref==3.5.1
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.2
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-freeze==0.6.4
django-js-asset==2.0.0
django-phonenumber-field==5.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
path==16.3.0
phonenumbers==8.12.28
Pillow==9.0.0
psycopg2-binary==2.9.3
python-telegram-bot==13.8
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
xmltodict==0.13.0
